addEmbedding

Given a Seurat seurat and a data frame where the rows correspond to cells, in the same order as in the Seurat seurat, add two columns giving coordinates in a dimensionality reduced space.

addEmbedding(seurat, df, reduction = "tsne", dim1 = 1, dim2 = 2)

Arguments

seurat
Seurat object, where dimensionality reduction has been applied, i.e. (after applying Seurat::RunPCA() or Seurat::RunTSNE() to the object).

df
Data frame with at least one column, "Cell", giving the cell ID

reduction
String specifying the dimensionality reduction to use, retrieves t-SNE by default. This should match the names of the elements of the list seurat@dr, so it will typically be one of "pca" or "tsne". Default: "tsne"

Examples

df <- data.frame(Cell = rownames(pbmc@meta.data),
                 Cluster = pbmc@meta.data$res.1)

addEmbedding(pbmc, df, reduction = "tsne")

#>              Cell Cluster        tSNE_1      tSNE_2
#> 1  ATGCCAGAACGACT       0  14.421908279   8.3360216
#> 2  CATGGCCTGTGCAT       0  16.723770814  11.2549069
#> 3  GAACCTGATGAACC       0   0.908030298  21.2748866
#> 4  TGACTGGATTCTCA       0 -12.271303667   8.6030550
#> 5  AGTCAGACTGCACA       0  16.201709739  11.9711820
#> 6  TCTGATACACGTGT       0  -1.503498217  26.5058488
#> 7  TGGTATCTAAACAG       0   3.196139126  26.8241782
#> 8  GCAGCTCTGTTTCT       0   3.041964808  25.6720057
#> 9  GATATAACACGCAT       0  -0.332479188  26.7644197
#> 10 AATGTTGACAGTCA       0   1.243669881  22.3781207
#> 11 AGGTCATGAGTGTC       0 -13.680731501  31.5102633
#> 12 AGAGATGATCTCGC       0  -6.293455087  28.1804780
#> 13 GGGTAACTCTAGTG       0 -12.790928570  32.0014392
#> 14 CATGAGACACGGGA       0  -4.331897062  26.7160161
#> 15 TACGCCACTCCGAA       1 -11.759544327  29.8174179
#> 16 CTAAACCTGTGCAT       1 -12.733136620  29.2751380
#> 17 GTAAGCACTCATTC       0  -8.248354895  29.6684085
#> 18 TTGGTACTGAATCC       0  -9.498695771  33.2646927
#> 19 CATCATACGGAGCA       1  -0.123897970   7.3137891
#> 20 TACATCACGCTAAC       0  -8.761689191  31.8784827
#> 21 TTACCATGAATCGC       1  11.267524801  -7.7502193
#> 22 ATAGGAGAAACAGA       1  10.830642792  -1.6648856
#> 23 GCGCACGACTTTAC       1  12.126009055  -4.1942573
#> 24 ACTCGCACGAAAGT       1  10.518837726  -2.8920338
#> 25 ATTACCTGCCTTAT       1  10.812888644  -5.3162386
#> 26 CCCAACTGCAATCG       1  11.641561769 -11.7115868
#> 27 AAATTCGAATCACG       1  11.029848830 -10.7859849
#> 28 CCATCCGATTCGCC       1   0.524525437   2.6680775
#> 29 TCCACTCTGAGCTT       1   2.842385660  -6.7659340
#> 30 CATCAGGATGCACA       0   1.765949332  13.7652065
#> 31 CTAAACCTCTGACA       2 -13.769229712  -0.9654148
#> 32 GATAGAGAAGGGTG       1   0.291315721  -1.4997801
#> 33 CTAACGGAACCGAT       2 -19.823177661  -7.6726257
#> 34 AGATATACCCGTAA       2 -16.412599106  -2.9953991
#> 35 TACTCTGAATCGAC       2 -21.290385096 -12.0934748
#> 36 GCGCATCTTGCTCC       2 -16.366509204  -1.5806007
#> 37 GTTGACGATATCGG       2 -19.154132197  -9.1918960
#> 38 ACAGGTACTGGTGT       2 -18.668051691 -10.6042361
#> 39 GGCATATGCTTATC       2 -19.815841785 -11.0518872
#> 40 CATTACACCAACTG       2 -19.525030175 -13.4461417
#> 41 TAGGGACTGAACTC       1  19.818541617  -0.7697982
#> 42 GCTCCATGAGAAGT       0  -3.227278554  17.5278759
#> 43 TACAATGATGCTAG       0  -1.563300862  17.8771146
#> 44 CTTCATGACCGAAT       0 -13.838199504   6.7204449
#> 45 CTGCCAACAGGAGC       0 -12.983243619   7.2517782
#> 46 TTGCATTGAGCTAC       0   2.966270149  13.5316278
#> 47 AAGCAAGAGCTTAG       2 -15.264921694  -2.2315300
#> 48 CGGCACGAACTCAG       0  -2.511786042  18.7598190
#> 49 GGTGGAGATTACTC       2 -13.718933190   4.5965815
#> 50 GGCCGATGTACTCT       0   1.611568997  27.1503153
#> 51 CGTAGCCTGTATGC       1  12.325827568  -9.9448526
#> 52 TGAGCTGAATGCTG       0  -0.003217834  15.5313851
#> 53 CCTATAACGAGACG       1   0.480240876   1.6813063
#> 54 ATAAGTTGGTACGT       1   0.516931555   4.9550165
#> 55 AAGCGACTTTGACG       0  -0.022527847   8.8484629
#> 56 ACCAGTGAATACCG       1  10.488968602   0.1365850
#> 57 ATTGCACTTGCTTT       2 -11.983627401  -0.6092689
#> 58 CTAGGTGATGGTTG       1  11.333136132   1.9259971
#> 59 GCACTAGACCTTTA       0  13.198956232   7.0309165
#> 60 CATGCGCTAGTCAC       1   4.133553319   2.7933328
#> 61 TTGAGGACTACGCA       1   2.220542569  -2.0733281
#> 62 ATACCACTCTAAGC       0   3.383287311  12.3017264
#> 63 CATATAGACTAAGC       1   4.004630434   0.7702884
#> 64 TTTAGCTGTACTCT       1  11.760652564   4.6291459
#> 65 GACATTCTCCACCT       1   3.443966235  -3.5010797
#> 66 ACGTGATGCCATGA       1   3.129281479  -0.1206174
#> 67 ATTGTAGATTCCCG       1   3.066266235  -4.8967971
#> 68 GATAGAGATCACGA       1  13.595787604  -1.5966398
#> 69 AATGCGTGGACGGA       1  18.760765827  -1.2000304
#> 70 GCGTAAACACGGTT       1  12.328389051   5.5517228
#> 71 ATTCAGCTCATTGG       3   6.441959517 -51.6852111
#> 72 GGCATATGGGGAGT       3   2.479364626 -50.9310316
#> 73 ATCATCTGACACCA       3   4.957757342 -54.9005134
#> 74 GTCATACTTCGCCT       3   5.936833536 -50.5698921
#> 75 TTACGTACGTTCAG       3   4.199273262 -50.3082081
#> 76 GAGTTGTGGTAGCT       3   6.349423075 -55.7263336
#> 77 GACGCTCTCTCTCG       3   6.577643443 -48.6447265
#> 78 AGTCTTACTTCGGA       3   3.468797905 -53.1112770
#> 79 GGAACACTTCAGAC       3   5.095322092 -52.2787746
#> 80 CTTGATTGATCTTC       3   4.808983374 -47.9329706