fetchData

Subset the seurat@data matrix by gene and cluster. Similar to Seurat::FetchData except it doesn't thrown an error if a gene is not found in the data, and is more limited.

fetchData(seurat, genes, clusters = NULL, return_cell = FALSE,
  return_cluster = FALSE, scaled = FALSE)

Arguments

seurat
Seurat object

genes
Genes to filter

clusters
(Optional) Vector, include only cells with these identities (e.g. cluster assignments). Searches in seurat@ident.

return_cell
Logical, whether or not to include a column with the cell ID. Default: FALSE

return_cluster
Logical, whether or not to include a column with the cluster. Default: FALSE

scaled
Logical, whether to fetch scaled data from seurat@scale.data. Default: FALSE.

Value

Expression matrix for genes specified

Examples

fetchData(pbmc, c("IL32", "MS4A1"))

#>       MS4A1     IL32
#> 1  0.000000 4.968821
#> 2  0.000000 0.000000
#> 3  0.000000 6.942623
#> 4  0.000000 6.447181
#> 5  0.000000 4.074201
#> 6  0.000000 0.000000
#> 7  0.000000 6.152201
#> 8  0.000000 6.034684
#> 9  0.000000 0.000000
#> 10 0.000000 5.707110
#> 11 5.779448 0.000000
#> 12 4.660719 0.000000
#> 13 5.984036 0.000000
#> 14 6.666083 0.000000
#> 15 5.313305 0.000000
#> 16 5.190573 0.000000
#> 17 6.106491 0.000000
#> 18 5.694729 0.000000
#> 19 5.537982 0.000000
#> 20 5.621232 0.000000
#> 21 0.000000 0.000000
#> 22 0.000000 0.000000
#> 23 0.000000 0.000000
#> 24 0.000000 0.000000
#> 25 3.117873 0.000000
#> 26 0.000000 0.000000
#> 27 0.000000 0.000000
#> 28 0.000000 0.000000
#> 29 0.000000 0.000000
#> 30 0.000000 0.000000
#> 31 0.000000 0.000000
#> 32 0.000000 5.164289
#> 33 0.000000 0.000000
#> 34 0.000000 5.592403
#> 35 0.000000 5.550671
#> 36 0.000000 0.000000
#> 37 0.000000 0.000000
#> 38 0.000000 0.000000
#> 39 0.000000 0.000000
#> 40 0.000000 5.405012
#> 41 0.000000 6.241874
#> 42 0.000000 5.888080
#> 43 0.000000 6.139805
#> 44 0.000000 0.000000
#> 45 0.000000 6.174727
#> 46 0.000000 4.576296
#> 47 0.000000 6.167916
#> 48 0.000000 6.614298
#> 49 0.000000 5.687374
#> 50 0.000000 4.625072
#> 51 0.000000 0.000000
#> 52 0.000000 0.000000
#> 53 0.000000 0.000000
#> 54 0.000000 0.000000
#> 55 0.000000 0.000000
#> 56 0.000000 0.000000
#> 57 0.000000 3.040721
#> 58 0.000000 3.461483
#> 59 0.000000 0.000000
#> 60 0.000000 0.000000
#> 61 0.000000 0.000000
#> 62 0.000000 0.000000
#> 63 0.000000 0.000000
#> 64 0.000000 0.000000
#> 65 0.000000 2.523157
#> 66 0.000000 0.000000
#> 67 0.000000 0.000000
#> 68 3.449600 0.000000
#> 69 0.000000 0.000000
#> 70 0.000000 0.000000
#> 71 0.000000 0.000000
#> 72 0.000000 0.000000
#> 73 0.000000 0.000000
#> 74 0.000000 0.000000
#> 75 0.000000 0.000000
#> 76 0.000000 0.000000
#> 77 0.000000 0.000000
#> 78 0.000000 0.000000
#> 79 0.000000 0.000000
#> 80 0.000000 3.782335

fetchData(pbmc, c("IL32"), c(1, 2))

#>        IL32
#> 1  0.000000
#> 2  0.000000
#> 3  0.000000
#> 4  0.000000
#> 5  0.000000
#> 6  0.000000
#> 7  0.000000
#> 8  0.000000
#> 9  0.000000
#> 10 0.000000
#> 11 0.000000
#> 12 0.000000
#> 13 0.000000
#> 14 5.164289
#> 15 0.000000
#> 16 5.592403
#> 17 5.550671
#> 18 0.000000
#> 19 0.000000
#> 20 0.000000
#> 21 0.000000
#> 22 5.405012
#> 23 6.241874
#> 24 6.167916
#> 25 5.687374
#> 26 0.000000
#> 27 0.000000
#> 28 0.000000
#> 29 0.000000
#> 30 3.040721
#> 31 3.461483
#> 32 0.000000
#> 33 0.000000
#> 34 0.000000
#> 35 0.000000
#> 36 2.523157
#> 37 0.000000
#> 38 0.000000
#> 39 0.000000
#> 40 0.000000
#> 41 0.000000

fetchData(pbmc, c("IL32", "MS4A1"), c(1, 2), return_cluster = TRUE, return_cell = TRUE)

#>              Cell Cluster    MS4A1     IL32
#> 1  TACGCCACTCCGAA       1 5.313305 0.000000
#> 2  CTAAACCTGTGCAT       1 5.190573 0.000000
#> 3  CATCATACGGAGCA       1 5.537982 0.000000
#> 4  TTACCATGAATCGC       1 0.000000 0.000000
#> 5  ATAGGAGAAACAGA       1 0.000000 0.000000
#> 6  GCGCACGACTTTAC       1 0.000000 0.000000
#> 7  ACTCGCACGAAAGT       1 0.000000 0.000000
#> 8  ATTACCTGCCTTAT       1 3.117873 0.000000
#> 9  CCCAACTGCAATCG       1 0.000000 0.000000
#> 10 AAATTCGAATCACG       1 0.000000 0.000000
#> 11 CCATCCGATTCGCC       1 0.000000 0.000000
#> 12 TCCACTCTGAGCTT       1 0.000000 0.000000
#> 13 CTAAACCTCTGACA       2 0.000000 0.000000
#> 14 GATAGAGAAGGGTG       1 0.000000 5.164289
#> 15 CTAACGGAACCGAT       2 0.000000 0.000000
#> 16 AGATATACCCGTAA       2 0.000000 5.592403
#> 17 TACTCTGAATCGAC       2 0.000000 5.550671
#> 18 GCGCATCTTGCTCC       2 0.000000 0.000000
#> 19 GTTGACGATATCGG       2 0.000000 0.000000
#> 20 ACAGGTACTGGTGT       2 0.000000 0.000000
#> 21 GGCATATGCTTATC       2 0.000000 0.000000
#> 22 CATTACACCAACTG       2 0.000000 5.405012
#> 23 TAGGGACTGAACTC       1 0.000000 6.241874
#> 24 AAGCAAGAGCTTAG       2 0.000000 6.167916
#> 25 GGTGGAGATTACTC       2 0.000000 5.687374
#> 26 CGTAGCCTGTATGC       1 0.000000 0.000000
#> 27 CCTATAACGAGACG       1 0.000000 0.000000
#> 28 ATAAGTTGGTACGT       1 0.000000 0.000000
#> 29 ACCAGTGAATACCG       1 0.000000 0.000000
#> 30 ATTGCACTTGCTTT       2 0.000000 3.040721
#> 31 CTAGGTGATGGTTG       1 0.000000 3.461483
#> 32 CATGCGCTAGTCAC       1 0.000000 0.000000
#> 33 TTGAGGACTACGCA       1 0.000000 0.000000
#> 34 CATATAGACTAAGC       1 0.000000 0.000000
#> 35 TTTAGCTGTACTCT       1 0.000000 0.000000
#> 36 GACATTCTCCACCT       1 0.000000 2.523157
#> 37 ACGTGATGCCATGA       1 0.000000 0.000000
#> 38 ATTGTAGATTCCCG       1 0.000000 0.000000
#> 39 GATAGAGATCACGA       1 3.449600 0.000000
#> 40 AATGCGTGGACGGA       1 0.000000 0.000000
#> 41 GCGTAAACACGGTT       1 0.000000 0.000000