Title

Title

getAllMarkers(df, labels, specify_clusters = NULL,
  output_graphs = FALSE, n_genes = 2, topn_markers = 10,
  graph_name = "Feature plot for cluster ", n_trees = NCOL(df) *
  n_genes, method = "RF", seurat)

Arguments

df
A dataframe with the genes as columns and cells as rows

labels
A dataframe that contains labels for the clustering solution. The dataframe should only have 1 column

specify_clusters
If users only want to extract markers for a subset of clusters, enter the subset here in the form of a vector

output_graphs
Indicates whether or not you wish to output the feature plots of the markers found to a pdf file

n_genes
The number of genes used in each decision tree

topn_markers
An integer indicating how many markers you wish to return for each cluster

graph_name
The name of the feature plot if one wishes to output them to a pdf file

n_trees
The number of trees one wishes to construct for each cluster

method
the method used for extracting markers, currently only the randomForest(RF) method is avalible

Value

A dataframe that contains the list of markers and their associated attributes

Examples

data <- t(pbmc@scale.data)
labels <- data.frame(as.numeric(pbmc@meta.data$res.1))
marker_list<-getAllMarkers(data, labels = labels)

#> [1] "retFrame constructed"
#> [1] "grouping worked properly"
#> [1] "ordering worked properly"
#> [1] "retFrame constructed"
#> [1] "grouping worked properly"
#> [1] "ordering worked properly"
#> [1] "retFrame constructed"
#> [1] "grouping worked properly"
#> [1] "ordering worked properly"
#> [1] "retFrame constructed"
#> [1] "grouping worked properly"
#> [1] "ordering worked properly"