meanGeneExpression

Calculate, for each cell, the mean expression of the given set of marker genes.

meanGeneExpression(seurat, genes)

Arguments

seurat
Seurat object

genes
String or character vector specifying gene(s) to use

Value

Data frame with two columns: "Cell" which specifies the cell ID obtained from colnames(seurat@data), and "Mean_marker_expression" which is the expression value for the marker gene, or mean if multiple genes were provided.

Examples

meanGeneExpression(pbmc, genes = c("IL32", "MS4A1"))

#>              Cell Mean_marker_expression
#> 1  ATGCCAGAACGACT               2.484410
#> 2  CATGGCCTGTGCAT               0.000000
#> 3  GAACCTGATGAACC               3.471312
#> 4  TGACTGGATTCTCA               3.223591
#> 5  AGTCAGACTGCACA               2.037100
#> 6  TCTGATACACGTGT               0.000000
#> 7  TGGTATCTAAACAG               3.076100
#> 8  GCAGCTCTGTTTCT               3.017342
#> 9  GATATAACACGCAT               0.000000
#> 10 AATGTTGACAGTCA               2.853555
#> 11 AGGTCATGAGTGTC               2.889724
#> 12 AGAGATGATCTCGC               2.330359
#> 13 GGGTAACTCTAGTG               2.992018
#> 14 CATGAGACACGGGA               3.333042
#> 15 TACGCCACTCCGAA               2.656653
#> 16 CTAAACCTGTGCAT               2.595287
#> 17 GTAAGCACTCATTC               3.053245
#> 18 TTGGTACTGAATCC               2.847365
#> 19 CATCATACGGAGCA               2.768991
#> 20 TACATCACGCTAAC               2.810616
#> 21 TTACCATGAATCGC               0.000000
#> 22 ATAGGAGAAACAGA               0.000000
#> 23 GCGCACGACTTTAC               0.000000
#> 24 ACTCGCACGAAAGT               0.000000
#> 25 ATTACCTGCCTTAT               1.558937
#> 26 CCCAACTGCAATCG               0.000000
#> 27 AAATTCGAATCACG               0.000000
#> 28 CCATCCGATTCGCC               0.000000
#> 29 TCCACTCTGAGCTT               0.000000
#> 30 CATCAGGATGCACA               0.000000
#> 31 CTAAACCTCTGACA               0.000000
#> 32 GATAGAGAAGGGTG               2.582144
#> 33 CTAACGGAACCGAT               0.000000
#> 34 AGATATACCCGTAA               2.796201
#> 35 TACTCTGAATCGAC               2.775336
#> 36 GCGCATCTTGCTCC               0.000000
#> 37 GTTGACGATATCGG               0.000000
#> 38 ACAGGTACTGGTGT               0.000000
#> 39 GGCATATGCTTATC               0.000000
#> 40 CATTACACCAACTG               2.702506
#> 41 TAGGGACTGAACTC               3.120937
#> 42 GCTCCATGAGAAGT               2.944040
#> 43 TACAATGATGCTAG               3.069902
#> 44 CTTCATGACCGAAT               0.000000
#> 45 CTGCCAACAGGAGC               3.087364
#> 46 TTGCATTGAGCTAC               2.288148
#> 47 AAGCAAGAGCTTAG               3.083958
#> 48 CGGCACGAACTCAG               3.307149
#> 49 GGTGGAGATTACTC               2.843687
#> 50 GGCCGATGTACTCT               2.312536
#> 51 CGTAGCCTGTATGC               0.000000
#> 52 TGAGCTGAATGCTG               0.000000
#> 53 CCTATAACGAGACG               0.000000
#> 54 ATAAGTTGGTACGT               0.000000
#> 55 AAGCGACTTTGACG               0.000000
#> 56 ACCAGTGAATACCG               0.000000
#> 57 ATTGCACTTGCTTT               1.520360
#> 58 CTAGGTGATGGTTG               1.730742
#> 59 GCACTAGACCTTTA               0.000000
#> 60 CATGCGCTAGTCAC               0.000000
#> 61 TTGAGGACTACGCA               0.000000
#> 62 ATACCACTCTAAGC               0.000000
#> 63 CATATAGACTAAGC               0.000000
#> 64 TTTAGCTGTACTCT               0.000000
#> 65 GACATTCTCCACCT               1.261578
#> 66 ACGTGATGCCATGA               0.000000
#> 67 ATTGTAGATTCCCG               0.000000
#> 68 GATAGAGATCACGA               1.724800
#> 69 AATGCGTGGACGGA               0.000000
#> 70 GCGTAAACACGGTT               0.000000
#> 71 ATTCAGCTCATTGG               0.000000
#> 72 GGCATATGGGGAGT               0.000000
#> 73 ATCATCTGACACCA               0.000000
#> 74 GTCATACTTCGCCT               0.000000
#> 75 TTACGTACGTTCAG               0.000000
#> 76 GAGTTGTGGTAGCT               0.000000
#> 77 GACGCTCTCTCTCG               0.000000
#> 78 AGTCTTACTTCGGA               0.000000
#> 79 GGAACACTTCAGAC               0.000000
#> 80 CTTGATTGATCTTC               1.891167