meanMarkerExpression

This function computes each cell's mean expression of each cluster's markers. Used for markerViolinPlot.

meanMarkerExpressionPerCell(seurat, markers, marker_col = "gene")

Arguments

seurat
Seurat object, whose expression values will be used

markers
Markers data frame, for the same or another Seurat object. The expression of markers in each cluster in this data frame will be calculated for each cluster in seurat.

Value

Data frame where the first two columns are "Cell" and "Cluster" (from seurat), and all remaining columns give the mean expression of markers identified for each cluster in markers

Examples

library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following object is masked from ‘package:testthat’:
#> 
#>     matches

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

# Using the same sample's marker:
meanMarkerExpressionPerCell(pbmc, markers_pbmc, "gene")

#>              Cell Cluster         0         1         2         3
#> 1  ATGCCAGAACGACT       0 1.9620095 1.0151715 0.9229822 1.4882036
#> 2  CATGGCCTGTGCAT       0 2.3624103 1.2569890 1.3528365 1.0867668
#> 3  GAACCTGATGAACC       0 2.4824218 1.2708467 0.5587361 1.0135824
#> 4  TGACTGGATTCTCA       0 2.6581993 1.0484821 0.8727627 1.5415155
#> 5  AGTCAGACTGCACA       0 2.2371453 1.0404154 1.2378454 1.3735021
#> 6  TCTGATACACGTGT       0 2.4229548 1.0191044 1.1041824 1.4093747
#> 7  TGGTATCTAAACAG       0 2.2647545 1.0283623 0.8021071 0.9895605
#> 8  GCAGCTCTGTTTCT       0 2.3852122 0.6375085 0.7839001 1.3055461
#> 9  GATATAACACGCAT       0 1.8612471 0.9713966 0.6092572 0.9487348
#> 10 AATGTTGACAGTCA       0 2.0459373 0.6726920 0.5433771 1.1835008
#> 11 AGGTCATGAGTGTC       0 0.5055272 0.7860749 0.7678067 0.8253274
#> 12 AGAGATGATCTCGC       0 1.2082666 1.2218916 1.3006091 1.2692123
#> 13 GGGTAACTCTAGTG       0 1.1350635 0.7474511 1.0552696 1.2036736
#> 14 CATGAGACACGGGA       0 0.5236486 0.7253361 0.6916279 0.8457214
#> 15 TACGCCACTCCGAA       1 0.9297920 0.9989216 1.0769440 1.3124356
#> 16 CTAAACCTGTGCAT       1 0.7250813 1.0453384 0.9794358 0.6580283
#> 17 GTAAGCACTCATTC       0 1.1708675 0.7861878 0.8812549 0.7560595
#> 18 TTGGTACTGAATCC       0 1.0870118 0.9300860 0.8945696 1.2544610
#> 19 CATCATACGGAGCA       1 1.1096844 0.9007748 1.1745053 0.7398573
#> 20 TACATCACGCTAAC       0 1.3665467 0.9740586 0.6502277 1.0392073
#> 21 TTACCATGAATCGC       1 2.1148161 2.7106021 0.8828995 1.3080543
#> 22 ATAGGAGAAACAGA       1 1.6306591 3.1628006 1.1895193 1.1978632
#> 23 GCGCACGACTTTAC       1 1.3850767 2.3656420 0.5251974 0.8225046
#> 24 ACTCGCACGAAAGT       1 1.7642553 2.6229904 0.7768852 1.1395130
#> 25 ATTACCTGCCTTAT       1 2.3821292 2.7437874 0.7722115 1.3957931
#> 26 CCCAACTGCAATCG       1 1.6291755 2.4817522 0.4998944 0.9774325
#> 27 AAATTCGAATCACG       1 2.1747420 2.6191155 0.8702053 1.1158092
#> 28 CCATCCGATTCGCC       1 1.7295463 2.1499197 0.7071736 0.8685511
#> 29 TCCACTCTGAGCTT       1 2.0022657 3.1312010 0.9628294 1.3728151
#> 30 CATCAGGATGCACA       0 2.0841263 2.9992996 1.0757341 1.8771719
#> 31 CTAAACCTCTGACA       2 1.8995999 0.9231039 3.1739210 1.1752963
#> 32 GATAGAGAAGGGTG       1 1.6719910 1.3040093 3.0898176 1.2354555
#> 33 CTAACGGAACCGAT       2 1.0047805 0.8290752 3.2862333 0.8685539
#> 34 AGATATACCCGTAA       2 1.9658540 1.0076273 3.2108509 1.2214919
#> 35 TACTCTGAATCGAC       2 1.4409857 0.9991682 2.8217216 0.5222615
#> 36 GCGCATCTTGCTCC       2 1.3120607 0.9829569 3.1431260 0.8666701
#> 37 GTTGACGATATCGG       2 1.9722686 1.3597328 3.6636082 1.2495518
#> 38 ACAGGTACTGGTGT       2 1.5681913 1.1981041 3.7314899 1.0892849
#> 39 GGCATATGCTTATC       2 1.8001399 0.8493369 3.4760961 1.2192140
#> 40 CATTACACCAACTG       2 1.8768900 0.8544381 3.4513014 1.2171510
#> 41 TAGGGACTGAACTC       1 1.9936002 1.0523481 2.5347753 1.2357729
#> 42 GCTCCATGAGAAGT       0 1.4180508 1.2427611 2.5530285 1.3895269
#> 43 TACAATGATGCTAG       0 1.8798690 0.9831174 1.5190663 1.0413507
#> 44 CTTCATGACCGAAT       0 1.8391942 0.5519590 1.5463178 1.3548650
#> 45 CTGCCAACAGGAGC       0 1.8800683 0.9923677 0.9507556 1.2940157
#> 46 TTGCATTGAGCTAC       0 1.4359172 0.9487827 1.3315580 1.3455303
#> 47 AAGCAAGAGCTTAG       2 1.1095572 0.8170908 2.7780854 1.1850883
#> 48 CGGCACGAACTCAG       0 2.1335493 0.8871076 2.2648022 1.4425958
#> 49 GGTGGAGATTACTC       2 1.2642318 0.9843162 2.1973019 1.1170138
#> 50 GGCCGATGTACTCT       0 1.9159130 0.8469066 1.4920259 1.2757057
#> 51 CGTAGCCTGTATGC       1 2.0958869 3.1403010 1.1311391 1.7515582
#> 52 TGAGCTGAATGCTG       0 2.2571650 2.7995726 1.3580815 1.9392769
#> 53 CCTATAACGAGACG       1 1.2222147 2.3594708 0.7575888 0.8491935
#> 54 ATAAGTTGGTACGT       1 1.2759473 2.1076508 0.8770911 0.7707009
#> 55 AAGCGACTTTGACG       0 1.6603312 2.8305003 1.2300028 1.2995569
#> 56 ACCAGTGAATACCG       1 1.6409718 2.8518136 0.8425639 1.2704907
#> 57 ATTGCACTTGCTTT       2 1.6372304 2.5903177 1.2040430 1.4259373
#> 58 CTAGGTGATGGTTG       1 1.6028336 2.8777927 0.9102645 1.3252146
#> 59 GCACTAGACCTTTA       0 1.8900687 2.9000835 1.2106504 1.5511662
#> 60 CATGCGCTAGTCAC       1 1.9904357 2.8572269 1.1019177 1.6752747
#> 61 TTGAGGACTACGCA       1 1.6678231 2.3971968 1.3793925 1.3719448
#> 62 ATACCACTCTAAGC       0 1.5740990 2.4759994 1.0739048 1.3125433
#> 63 CATATAGACTAAGC       1 1.6690427 2.5613321 1.1711037 1.2772649
#> 64 TTTAGCTGTACTCT       1 2.2610955 2.8774915 1.1994113 1.4454629
#> 65 GACATTCTCCACCT       1 1.9624785 2.6037289 1.6514179 1.5153782
#> 66 ACGTGATGCCATGA       1 2.1702266 3.0847040 1.4619390 1.5439810
#> 67 ATTGTAGATTCCCG       1 1.9048612 2.8551012 1.3372740 1.5517786
#> 68 GATAGAGATCACGA       1 1.9273085 2.9429186 1.0061882 1.1591394
#> 69 AATGCGTGGACGGA       1 2.0026104 2.7780448 1.1020487 1.2818843
#> 70 GCGTAAACACGGTT       1 1.7325230 2.6216423 1.1832652 1.4643800
#> 71 ATTCAGCTCATTGG       3 0.8628656 0.8440529 0.5642226 3.2839664
#> 72 GGCATATGGGGAGT       3 0.5796475 0.5828446 0.3533345 2.6592873
#> 73 ATCATCTGACACCA       3 0.7051697 0.6215072 0.4501248 3.3070859
#> 74 GTCATACTTCGCCT       3 0.5980868 0.5366269 0.3886558 3.0448698
#> 75 TTACGTACGTTCAG       3 0.9635830 0.8321858 0.5543588 3.1541527
#> 76 GAGTTGTGGTAGCT       3 0.9642359 0.7875461 0.6132402 3.3056136
#> 77 GACGCTCTCTCTCG       3 0.6612884 0.5760890 0.3601758 2.8475710
#> 78 AGTCTTACTTCGGA       3 0.5855287 0.4038641 0.4628059 2.7701330
#> 79 GGAACACTTCAGAC       3 0.6326417 0.5764848 0.4060478 2.9264419
#> 80 CTTGATTGATCTTC       3 1.8330036 2.8231538 0.9663513 2.3550694

# Change the name of the clusters in the markers df, as if it were
# from a different sample where the clusters are A, B, C, D:
markers2 <- mutate(markers_pbmc, cluster = recode(
    cluster, `0` = "A", `1` = "B", `2` = "C", `3` = "D"))

meanMarkerExpressionPerCell(pbmc, markers2, "gene")

#>              Cell Cluster         A         B         C         D
#> 1  ATGCCAGAACGACT       0 1.9620095 1.0151715 0.9229822 1.4882036
#> 2  CATGGCCTGTGCAT       0 2.3624103 1.2569890 1.3528365 1.0867668
#> 3  GAACCTGATGAACC       0 2.4824218 1.2708467 0.5587361 1.0135824
#> 4  TGACTGGATTCTCA       0 2.6581993 1.0484821 0.8727627 1.5415155
#> 5  AGTCAGACTGCACA       0 2.2371453 1.0404154 1.2378454 1.3735021
#> 6  TCTGATACACGTGT       0 2.4229548 1.0191044 1.1041824 1.4093747
#> 7  TGGTATCTAAACAG       0 2.2647545 1.0283623 0.8021071 0.9895605
#> 8  GCAGCTCTGTTTCT       0 2.3852122 0.6375085 0.7839001 1.3055461
#> 9  GATATAACACGCAT       0 1.8612471 0.9713966 0.6092572 0.9487348
#> 10 AATGTTGACAGTCA       0 2.0459373 0.6726920 0.5433771 1.1835008
#> 11 AGGTCATGAGTGTC       0 0.5055272 0.7860749 0.7678067 0.8253274
#> 12 AGAGATGATCTCGC       0 1.2082666 1.2218916 1.3006091 1.2692123
#> 13 GGGTAACTCTAGTG       0 1.1350635 0.7474511 1.0552696 1.2036736
#> 14 CATGAGACACGGGA       0 0.5236486 0.7253361 0.6916279 0.8457214
#> 15 TACGCCACTCCGAA       1 0.9297920 0.9989216 1.0769440 1.3124356
#> 16 CTAAACCTGTGCAT       1 0.7250813 1.0453384 0.9794358 0.6580283
#> 17 GTAAGCACTCATTC       0 1.1708675 0.7861878 0.8812549 0.7560595
#> 18 TTGGTACTGAATCC       0 1.0870118 0.9300860 0.8945696 1.2544610
#> 19 CATCATACGGAGCA       1 1.1096844 0.9007748 1.1745053 0.7398573
#> 20 TACATCACGCTAAC       0 1.3665467 0.9740586 0.6502277 1.0392073
#> 21 TTACCATGAATCGC       1 2.1148161 2.7106021 0.8828995 1.3080543
#> 22 ATAGGAGAAACAGA       1 1.6306591 3.1628006 1.1895193 1.1978632
#> 23 GCGCACGACTTTAC       1 1.3850767 2.3656420 0.5251974 0.8225046
#> 24 ACTCGCACGAAAGT       1 1.7642553 2.6229904 0.7768852 1.1395130
#> 25 ATTACCTGCCTTAT       1 2.3821292 2.7437874 0.7722115 1.3957931
#> 26 CCCAACTGCAATCG       1 1.6291755 2.4817522 0.4998944 0.9774325
#> 27 AAATTCGAATCACG       1 2.1747420 2.6191155 0.8702053 1.1158092
#> 28 CCATCCGATTCGCC       1 1.7295463 2.1499197 0.7071736 0.8685511
#> 29 TCCACTCTGAGCTT       1 2.0022657 3.1312010 0.9628294 1.3728151
#> 30 CATCAGGATGCACA       0 2.0841263 2.9992996 1.0757341 1.8771719
#> 31 CTAAACCTCTGACA       2 1.8995999 0.9231039 3.1739210 1.1752963
#> 32 GATAGAGAAGGGTG       1 1.6719910 1.3040093 3.0898176 1.2354555
#> 33 CTAACGGAACCGAT       2 1.0047805 0.8290752 3.2862333 0.8685539
#> 34 AGATATACCCGTAA       2 1.9658540 1.0076273 3.2108509 1.2214919
#> 35 TACTCTGAATCGAC       2 1.4409857 0.9991682 2.8217216 0.5222615
#> 36 GCGCATCTTGCTCC       2 1.3120607 0.9829569 3.1431260 0.8666701
#> 37 GTTGACGATATCGG       2 1.9722686 1.3597328 3.6636082 1.2495518
#> 38 ACAGGTACTGGTGT       2 1.5681913 1.1981041 3.7314899 1.0892849
#> 39 GGCATATGCTTATC       2 1.8001399 0.8493369 3.4760961 1.2192140
#> 40 CATTACACCAACTG       2 1.8768900 0.8544381 3.4513014 1.2171510
#> 41 TAGGGACTGAACTC       1 1.9936002 1.0523481 2.5347753 1.2357729
#> 42 GCTCCATGAGAAGT       0 1.4180508 1.2427611 2.5530285 1.3895269
#> 43 TACAATGATGCTAG       0 1.8798690 0.9831174 1.5190663 1.0413507
#> 44 CTTCATGACCGAAT       0 1.8391942 0.5519590 1.5463178 1.3548650
#> 45 CTGCCAACAGGAGC       0 1.8800683 0.9923677 0.9507556 1.2940157
#> 46 TTGCATTGAGCTAC       0 1.4359172 0.9487827 1.3315580 1.3455303
#> 47 AAGCAAGAGCTTAG       2 1.1095572 0.8170908 2.7780854 1.1850883
#> 48 CGGCACGAACTCAG       0 2.1335493 0.8871076 2.2648022 1.4425958
#> 49 GGTGGAGATTACTC       2 1.2642318 0.9843162 2.1973019 1.1170138
#> 50 GGCCGATGTACTCT       0 1.9159130 0.8469066 1.4920259 1.2757057
#> 51 CGTAGCCTGTATGC       1 2.0958869 3.1403010 1.1311391 1.7515582
#> 52 TGAGCTGAATGCTG       0 2.2571650 2.7995726 1.3580815 1.9392769
#> 53 CCTATAACGAGACG       1 1.2222147 2.3594708 0.7575888 0.8491935
#> 54 ATAAGTTGGTACGT       1 1.2759473 2.1076508 0.8770911 0.7707009
#> 55 AAGCGACTTTGACG       0 1.6603312 2.8305003 1.2300028 1.2995569
#> 56 ACCAGTGAATACCG       1 1.6409718 2.8518136 0.8425639 1.2704907
#> 57 ATTGCACTTGCTTT       2 1.6372304 2.5903177 1.2040430 1.4259373
#> 58 CTAGGTGATGGTTG       1 1.6028336 2.8777927 0.9102645 1.3252146
#> 59 GCACTAGACCTTTA       0 1.8900687 2.9000835 1.2106504 1.5511662
#> 60 CATGCGCTAGTCAC       1 1.9904357 2.8572269 1.1019177 1.6752747
#> 61 TTGAGGACTACGCA       1 1.6678231 2.3971968 1.3793925 1.3719448
#> 62 ATACCACTCTAAGC       0 1.5740990 2.4759994 1.0739048 1.3125433
#> 63 CATATAGACTAAGC       1 1.6690427 2.5613321 1.1711037 1.2772649
#> 64 TTTAGCTGTACTCT       1 2.2610955 2.8774915 1.1994113 1.4454629
#> 65 GACATTCTCCACCT       1 1.9624785 2.6037289 1.6514179 1.5153782
#> 66 ACGTGATGCCATGA       1 2.1702266 3.0847040 1.4619390 1.5439810
#> 67 ATTGTAGATTCCCG       1 1.9048612 2.8551012 1.3372740 1.5517786
#> 68 GATAGAGATCACGA       1 1.9273085 2.9429186 1.0061882 1.1591394
#> 69 AATGCGTGGACGGA       1 2.0026104 2.7780448 1.1020487 1.2818843
#> 70 GCGTAAACACGGTT       1 1.7325230 2.6216423 1.1832652 1.4643800
#> 71 ATTCAGCTCATTGG       3 0.8628656 0.8440529 0.5642226 3.2839664
#> 72 GGCATATGGGGAGT       3 0.5796475 0.5828446 0.3533345 2.6592873
#> 73 ATCATCTGACACCA       3 0.7051697 0.6215072 0.4501248 3.3070859
#> 74 GTCATACTTCGCCT       3 0.5980868 0.5366269 0.3886558 3.0448698
#> 75 TTACGTACGTTCAG       3 0.9635830 0.8321858 0.5543588 3.1541527
#> 76 GAGTTGTGGTAGCT       3 0.9642359 0.7875461 0.6132402 3.3056136
#> 77 GACGCTCTCTCTCG       3 0.6612884 0.5760890 0.3601758 2.8475710
#> 78 AGTCTTACTTCGGA       3 0.5855287 0.4038641 0.4628059 2.7701330
#> 79 GGAACACTTCAGAC       3 0.6326417 0.5764848 0.4060478 2.9264419
#> 80 CTTGATTGATCTTC       3 1.8330036 2.8231538 0.9663513 2.3550694